pyproject.toml
# pyproject.toml
[project]
name = "mental-health-multimodal"
version = "0.1.0"
description = "Multimodal mental health risk assessment using AI"
authors = [
    {name = "Your Name", email = "your.email@example.com"},
]
readme = "README.md"
requires-python = ">=3.8"
license = {text = "Apache-2.0"}

dependencies = [
    "torch>=2.0.0",
    "transformers>=4.30.0",
    "torchaudio>=2.0.0",
    "librosa>=0.10.0",
    "fastapi>=0.100.0",
    "uvicorn>=0.23.0",
    "pydantic>=2.0.0",
    "scikit-learn>=1.2.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "python-multipart>=0.0.6",
    "aiofiles>=23.0.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.3.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "pre-commit>=3.3.0",
    "mlflow>=2.7.0",
    "wandb>=0.15.0",
]

[build-system]
requires = ["setuptools>=61.0.0", "wheel"]
build-backend = "setuptools.build_meta"